requirements.txt
# Stock Screener Dependencies

# Core
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0  # Progress bars for better UX

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
lxml>=4.9.0  # Required for pandas.read_html()
html5lib>=1.1  # Alternative HTML parser for pandas

# Yahoo Finance
yfinance>=0.2.28

# Optional: For enhanced functionality
# openpyxl>=3.1.0  # For Excel export
# ta>=0.11.0  # Technical analysis
# matplotlib>=3.7.0  # For visualization
# seaborn>=0.12.0  # For visualization
